Added cudnn functions
- Added NPP class for image resize, normalize, and converting to CHW format.
- Added CNNlayer class forward pass with cudnn functions.
- Added batching.
- Moved the raw cuda implementation to simple folder.
