+++ .vscode/settings.json
@@ -65,6 +65,15 @@
         "list": "cpp",
         "xfacet": "cpp",
         "xlocale": "cpp",
-        "xlocinfo": "cpp"
+        "xlocinfo": "cpp",
+        "array": "cpp",
+        "codecvt": "cpp",
+        "complex": "cpp",
+        "map": "cpp",
+        "mutex": "cpp",
+        "set": "cpp",
+        "stop_token": "cpp",
+        "thread": "cpp",
+        "xtree": "cpp"
     }
 }
