+++ advanced/lib/kernel_init.h
@@ -1,0 +1,3 @@
+#include <curand_kernel.h>
+
+__global__ void initializeWeights(float* weights, int size, unsigned long long seed, float min, float max);
